Move CSS and JS to external files
